Fix test framework and add Themes screen
Test Framework Fixes:
- Fix coroutine wait timing with wait_until_frame in promise_type
- Add missing UI update systems to test runner (register_before/after_ui_updates)
- Fix mouse position interception (enable GetMousePosition_Test wrapper)
- Fix click simulation to apply translation modifiers for proper positioning
- Fix double key processing by removing duplicate input_injector::set_key_down
- Add mouse release simulation for proper click detection
- Move reset_frame to end of TestSystem::once for proper input state

Test Updates:
- Update TabbingTest to match auto-focus behavior
- Adjust wait frame counts for proper timing

New Features:
- Add Themes.h screen for real-time theme switching demo with shadows

Warning Fixes:
- Fix size_t to EntityID conversions in screen files
- Fix unused variable in ExampleAccessibility.h
- Add missing switch default cases in Themes.h

+++ src/systems/screens/Buttons.h
@@ -198,7 +198,7 @@ struct ButtonsGallery : afterhours::System<UIContext<InputAction>> {
       else if (i == group_labels.size() - 1) corners = corners.sharp(TOP_LEFT).sharp(BOTTOM_LEFT);
       else corners = corners.all_sharp();
       
-      if (button(context, mk(row3.ent(), 1 + i),
+      if (button(context, mk(row3.ent(), 1 + static_cast<int>(i)),
                  ComponentConfig{}
                      .with_label(std::string(group_labels[i]))
                      .with_size(ComponentSize{pixels(90), pixels(45)})
